+++ src/twalk.h
@@ -303,7 +303,7 @@ class twalk {
 		free_vector(mapx);
 		free_vector(y);
 		free_vector(yp);
-		delete[] phi; // tmp MB
+		delete phi; // was changed by MB to delete[] phi;
 	}
 
 		/* select a kernel accordingly with krl_probs */
@@ -312,7 +312,7 @@ class twalk {
 
 		double aux = Un01();
 
-        kernel *k=NULL;
+        kernel *k=NULL; // jac's version says kernel *k;
 
 
         if(aux<0.0000 )                         //kernel K0, not in use
@@ -526,14 +526,14 @@ int one_move() {
 
 
 /***  backwards compatibility !!!!!!!!!!! ***/
-int simulation(int Tr1, char *filename, const char* op="wt") {
+int simulation(int Tr1, char *filename, const char* op="wt") { // jac's rplum says const char *filename, char* op="wt"
 
 	simulation( Tr1, filename, op="wt", (int) 0, NULL, NULL);
     return 1;
 }
 
 /* Here is the implementation of the central part of the algorithm */
-int simulation(unsigned long long Tr1, char *filename, const char *op="wt", int save_every1=1, double *xx=NULL, double *xxp=NULL, int silent=0) {
+int simulation(unsigned long long Tr1, char *filename, const char *op="wt", int save_every1=1, double *xx=NULL, double *xxp=NULL, int silent=0) { // jac's twalk.h has no silent=0
 
 
 	FILE *fptr;
@@ -542,7 +542,7 @@ int simulation(unsigned long long Tr1, char *filename, const char *op="wt", int 
     long sec=time(NULL); //beginning of the twalk
     time_t temp = sec;
     if (silent == 0)
-		Rprintf("twalk: %12lu iterations to run", Tr1, ctime(&temp));
+        Rprintf("twalk: %12lu iterations to run, %s", Tr1, ctime(&temp));
 
     // ----- --- Initialization --- -----
 
@@ -615,7 +615,7 @@ int simulation(unsigned long long Tr1, char *filename, const char *op="wt", int 
 
 		Rprintf("BUFSIZ is %d, optimal block size changed to %ld\n", BUFSIZ, st_blksize);
 #endif
-        
+
 		fver_vector(fptr, x, n);
         fprintf(fptr, "\t %f", U); // was %lf MB
 
@@ -724,8 +724,6 @@ int simulation(unsigned long long Tr1, char *filename, const char *op="wt", int 
 		sec = time(NULL);
         temp = sec;
 		if (silent == 0) {
-            //Rprintf("\nOh, la semilla esta en %lu \n", GetSeed()); // tmp MB Oct 2020	
-
             Rprintf("twalk: Finished, %4.1f%% of moved pars per iteration (ratio %f/%lu). Output in file %s,\n      %s\n",
                  100.0*(acc/(double) Tr1), acc, Tr1, filename, ctime(&temp));
         }